refactor: 스탬프 컴포넌트 분리
